// src/types.ts
export type Theme = 'midnight' | 'emerald' | 'ocean' | 'rose' | 'light' | 'gold' | 'system';

export type AppSize = 'small' | 'medium' | 'large';